refactor: move the device info into its own component
