+++ frontend/src/components/fleet/devices/storage-table/StorageTable.tsx
@@ -28,6 +28,7 @@ import type {
 
 import Result from "@/components/ui/result/Result";
 import Table, { createColumnHelper } from "@/components/ui/table/Table";
+import formatBytes from "@/utils/formatBytes";
 
 // We use graphql fields below in columns configuration
 /* eslint-disable relay/unused-fields */
@@ -44,18 +45,6 @@ const STORAGE_TABLE_FRAGMENT = graphql`
 type StorageUnit = NonNullable<
   StorageTable_storageUsage$data["storageUsage"]
 >[number];
-
-const formatBytes = (bytes: number, decimals = 2) => {
-  if (bytes === 0) return "0 B";
-
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ["B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB"];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + " " + sizes[i];
-};
 
 const columnHelper = createColumnHelper<StorageUnit>();
 const columns = [
