+++ frontend/src/i18n/langs/en.json
@@ -1213,6 +1213,53 @@
   "components.fleet.devices.connection-status.ConnectionStatus.statusDisconnected": {
     "defaultMessage": "Offline"
   },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.checkMyDevice": {
+    "defaultMessage": "Check my Device"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.connectionStatus": {
+    "defaultMessage": "Connection"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.deviceId": {
+    "defaultMessage": "Device ID"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.groups": {
+    "defaultMessage": "Groups"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.hardwareType": {
+    "defaultMessage": "Hardware Type"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.lastSeen": {
+    "defaultMessage": "Last seen"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.name": {
+    "defaultMessage": "Name"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.openRemoteTerminalErrorFeedback": {
+    "defaultMessage": "Could not access the remote terminal, please try again.",
+    "description": "Feedback for unknown error while opening a remote terminal session"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.partNumber": {
+    "defaultMessage": "Part Number"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.remoteTerminal.label": {
+    "defaultMessage": "Remote Terminal"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.remoteTerminal.openTerminalButton": {
+    "defaultMessage": "Open"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.serialNumber": {
+    "defaultMessage": "Serial Number"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.systemModel": {
+    "defaultMessage": "System Model"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.tags": {
+    "defaultMessage": "Tags"
+  },
+  "components.fleet.devices.device-info-card.DeviceInfoCard.updateDeviceErrorFeedback": {
+    "defaultMessage": "Could not update the device, please try again.",
+    "description": "Feedback for unknown error while updating a device"
+  },
   "components.fleet.devices.devices-table.DevicesTable.deviceIdTitle": {
     "defaultMessage": "Device ID",
     "description": "Title for the Device ID column of the devices table"
@@ -3669,58 +3716,11 @@
   "pages.Deployments.title": {
     "defaultMessage": "Deployments"
   },
-  "pages.Device.checkMyDevice": {
-    "defaultMessage": "Check my Device"
-  },
-  "pages.Device.connectionStatus": {
-    "defaultMessage": "Connection"
-  },
-  "pages.Device.deviceId": {
-    "defaultMessage": "Device ID"
-  },
   "pages.Device.deviceNotFound.message": {
     "defaultMessage": "Return to the device list."
   },
   "pages.Device.deviceNotFound.title": {
     "defaultMessage": "Device not found."
-  },
-  "pages.Device.groups": {
-    "defaultMessage": "Groups"
-  },
-  "pages.Device.hardwareType": {
-    "defaultMessage": "Hardware Type"
-  },
-  "pages.Device.lastSeen": {
-    "defaultMessage": "Last seen"
-  },
-  "pages.Device.name": {
-    "defaultMessage": "Name"
-  },
-  "pages.Device.openRemoteTerminalErrorFeedback": {
-    "defaultMessage": "Could not access the remote terminal, please try again.",
-    "description": "Feedback for unknown error while opening a remote terminal session"
-  },
-  "pages.Device.partNumber": {
-    "defaultMessage": "Part Number"
-  },
-  "pages.Device.remoteTerminal.label": {
-    "defaultMessage": "Remote Terminal"
-  },
-  "pages.Device.remoteTerminal.openTerminalButton": {
-    "defaultMessage": "Open"
-  },
-  "pages.Device.serialNumber": {
-    "defaultMessage": "Serial Number"
-  },
-  "pages.Device.systemModel": {
-    "defaultMessage": "System Model"
-  },
-  "pages.Device.tags": {
-    "defaultMessage": "Tags"
-  },
-  "pages.Device.updateDeviceErrorFeedback": {
-    "defaultMessage": "Could not update the device, please try again.",
-    "description": "Feedback for unknown error while updating a device"
   },
   "pages.DeviceGroup.deleteModal.description": {
     "defaultMessage": "This action cannot be undone. This will permanently delete the Group <bold>{deviceGroup}</bold>.",
